@@ -1,54 +1,30 @@
 {
-  "homepage": "/",
-  "name": "bottlecrm",
+  "name": "workspace",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@babel/core": "^7.15.8",
-    "@babel/preset-react": "^7.14.5",
-    "@ckeditor/ckeditor5-build-classic": "^31.0.0",
-    "@ckeditor/ckeditor5-react": "^3.0.3",
-    "@sentry/react": "^6.13.3",
-    "@sentry/tracing": "^6.13.3",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "antd": "^4.16.13",
-    "axios": "^0.24.0",
-    "babel-polyfill": "^6.26.0",
-    "framer-motion": "^4.1.17",
-    "html-react-parser": "^1.4.0",
-    "html-webpack-plugin": "^4.5.2",
-    "jodit-react": "^1.1.10",
-    "moment": "^2.29.1",
-    "react": "^17.0.2",
-    "react-beautiful-dnd": "^13.1.0",
-    "react-custom-scrollbars-2": "^4.4.0",
-    "react-dom": "^17.0.2",
-    "react-quill": "^2.0.0-beta.2",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
-    "redux": "^4.1.2",
-    "redux-saga": "^1.1.3",
-    "web-vitals": "^1.1.1",
-    "webpack": "4.44.2",
-    "webpack-cli": "^4.6.0",
-    "webpack-dev-server": "3.11.1"
+    "@emotion/react": "^11.10.0",
+    "@emotion/styled": "^11.10.0",
+    "@mui/icons-material": "^5.8.4",
+    "@mui/material": "^5.10.0",
+    "@mui/styles": "^5.9.3",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.4.3",
+    "mui-quill": "0.1.0-alpha.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-phone-number-input": "^3.2.7",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "5.0.0",
+    "validator": "^13.7.0",
+    "web-vitals": "^2.1.0"
   },
   "scripts": {
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
     "start": "react-scripts start",
-    "serve": "webpack serve --config webpack.config.js --progress --profile --watch --content-base src/",
     "build": "react-scripts build",
-    "eslint": "eslint src/**/*.js",
-    "webpack-serve": "webpack-dev-server --mode development",
-    "webpack-server": "npm run serve && node dist/server.js",
-    "format": "prettier --write \"src/**/*.js\"",
-    "eslint-fix": "eslint --fix \"src/**/*.js\"",
-    "webpack-start": "webpack serve --mode development",
-    "webpack-build": "webpack --mode production"
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
@@ -67,8 +43,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "devDependencies": {
-    "cypress": "^8.7.0"
   }
 }
